pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdna-miner"
version = "0.1.0"
description = "rDNA-miner: extract, assemble, and classify rDNA from long reads"
authors = [
    { name="FDB" },
    { name="AR" },
    { name="DG" }
]
dependencies = [
    "biopython>=1.79",
    "pandas>=1.3",
    "typer>=0.7",
    "numpy>=1.21"
]

[project.urls]
"Source" = "https://github.com/FDBoever/rDNA-miner"

[project.scripts]
rdna-miner = "rdna_miner.cli:app"